vector2d/2f for more customizability i got the idea from minecraft

+++ NeedForCheese/world.h
@@ -15,6 +15,6 @@ const int SCALED_WIDTH = WIDTH / MET2PIX; // The width of the window, in meters.
 const int SCALED_HEIGHT = HEIGHT / MET2PIX; // The height of the window, in meters.
 
 //1 rad x 180/pi = 57,296deg
-const float RAD2DEG = 180 / M_PI; // Value to convert radians to degrees
+const double RAD2DEG = 180 / M_PI; // Value to convert radians to degrees
 
 #endif // !WORLD_H
